inital

+++ app/Services/FileService.php
@@ -2,21 +2,30 @@
 
 
 namespace App\Services;
-use Image;
+
+
 
 class FileService
 {
     public function storeFile($data)
     {
-        $imageData = $data;
-        $fileName = uniqid().'.'.explode('/',
-                explode(':', substr($imageData, 0, strpos($imageData, ';')))[1])[1];
-        Image::make($imageData)->save('image/'.$fileName);
+
+        $file = $data->file('image');
+        $fileName = time() . '.' . $file->getClientOriginalExtension();
+        $file->move(public_path() . '/image', $fileName);
         return 'image/' . $fileName;
+
+//        $fileName = uniqid().'.'.explode('/',
+//                explode(':', substr($imageData, 0, strpos($imageData, ';')))[1])[1];
+//        Image::make($imageData)->save('image/'.$fileName);
+//        return 'image/' . $fileName;
     }
 
     public function removeFile($filePath)
     {
-        unlink(public_path("$filePath"));
+        if(file_exists(public_path("$filePath"))){
+            unlink(public_path("$filePath"));
+        }
+
     }
 }
